- Added a wrapper class for engine to allow lazy loading - Created pydantic models for validation - Separated the engine from the main fastapi app
